more double to float stuff

+++ src/MyMediaLite/Eval/FoldIn.cs
@@ -66,10 +66,10 @@ namespace MyMediaLite.Eval
 			cbd  = cbd / rating_count;
 
 			var result = new RatingPredictionEvaluationResults();
-			result["RMSE"] = rmse;
-			result["MAE"]  = mae;
-			result["NMAE"] = mae / (recommender.MaxRating - recommender.MinRating);
-			result["CBD"]  = cbd;
+			result["RMSE"] = (float) rmse;
+			result["MAE"]  = (float) mae;
+			result["NMAE"] = (float) mae / (recommender.MaxRating - recommender.MinRating);
+			result["CBD"]  = (float) cbd;
 			return result;
 		}
 
@@ -120,10 +120,10 @@ namespace MyMediaLite.Eval
 			cbd  = cbd / rating_count;
 
 			var result = new RatingPredictionEvaluationResults();
-			result["RMSE"] = rmse;
-			result["MAE"]  = mae;
-			result["NMAE"] = mae / (recommender.MaxRating - recommender.MinRating);
-			result["CBD"]  = cbd;
+			result["RMSE"] = (float) rmse;
+			result["MAE"]  = (float) mae;
+			result["NMAE"] = (float) mae / (recommender.MaxRating - recommender.MinRating);
+			result["CBD"]  = (float) cbd;
 			return result;
 		}
 
@@ -177,10 +177,10 @@ namespace MyMediaLite.Eval
 			cbd  = cbd / rating_count;
 
 			var result = new RatingPredictionEvaluationResults();
-			result["RMSE"] = rmse;
-			result["MAE"]  = mae;
-			result["NMAE"] = mae / (recommender.MaxRating - recommender.MinRating);
-			result["CBD"]  = cbd;
+			result["RMSE"] = (float) rmse;
+			result["MAE"]  = (float) mae;
+			result["NMAE"] = (float) mae / (recommender.MaxRating - recommender.MinRating);
+			result["CBD"]  = (float) cbd;
 			return result;
 		}
 	}
