+++ src/MyMediaLite/Eval/ItemRecommendationEvaluationResults.cs
@@ -26,7 +26,7 @@ namespace MyMediaLite.Eval
 	/// <remarks>
 	/// This class is basically a Dictionary with a custom-made ToString() method.
 	/// </remarks>
-	public class ItemRecommendationEvaluationResults : Dictionary<string, double>
+	public class ItemRecommendationEvaluationResults : Dictionary<string, float>
 	{
 		/// <summary>default constructor</summary>
 		public ItemRecommendationEvaluationResults()
